requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-dht==4.0.6
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
annotated-types==0.7.0
anyio==4.6.2.post1
binho-host-adapter==0.1.6
blinker==1.9.0
board==1.0
click==8.1.7
dht11==0.1.0
fastapi==0.115.6
Flask==3.1.0
Flask-Cors==5.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pydantic==2.10.3
pydantic_core==2.27.1
pyftdi==0.56.0
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
schedule==1.2.2
sniffio==1.3.1
starlette==0.41.3
sysv-ipc==1.1.0
typing_extensions==4.12.2
uvicorn==0.32.1
Werkzeug==3.1.3
